refactor: move common types and guard functions to common.ts
